requirements.txt
# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9

# ML Core
torch==2.3.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
xgboost==2.0.3
lightgbm==4.3.0
imbalanced-learn==0.12.3
shap==0.45.1

# Feature Engineering
scipy==1.13.0
haversine==2.8.0

# Streaming
kafka-python==2.0.2
pyspark==3.5.1

# LLM Agent
openai==1.30.1
langchain==0.2.1
langchain-openai==0.1.8
langchain-core==0.2.2

# Observability
prometheus-client==0.20.0
structlog==24.1.0

# Storage
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
tenacity==8.3.0
tqdm==4.66.4
joblib==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
pytest-mock==3.14.0

# Linting
black==24.10.0
ruff==0.4.4
